+++ src/components/Tilemap/index.ts
@@ -1,0 +1,3 @@
+export { Tilemap } from './Tilemap';
+export { Tiles } from './Tiles';
+export { useTiles } from './useTiles';