+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.1.0
 
+- Updated `Tilemap` to support property based tile data.
 - Added `useTime` hook, giving the number of ms since the update hook started.
 - Added `randi` utility function for generating random integers.
 - Improved performance by removing invisible elements from the render tree.